Remove some C-style casts and double bass drum output value

+++ src/adplug/cmf.cpp
@@ -278,7 +278,7 @@ bool CcmfPlayer::update() {
             // 8192 is middle/off, 0 is -2 semitones, 16384 is +2 semitones
             this->chMIDI[iChannel].iPitchbend = iValue;
             AdPlug_LogWrite("CMF: Channel %d pitchbent to %d (%+.2f)\n", iChannel + 1,
-                            iValue, (float)(iValue - 8192) / 8192);
+                            iValue, (iValue - 8192.0f) / 8192);
             break;
         }
         case 0xF0: // System message (arbitrary data bytes)
@@ -522,10 +522,10 @@ void CcmfPlayer::cmfNoteOn(uint8_t iChannel, uint8_t iNote, uint8_t iVelocity) {
     //if (iBlock > 7) iBlock = 7; // don't want to go out of range
 
     double d =
-            pow(2, ((double) iNote + ((this->chMIDI[iChannel].iPitchbend - 8192) /
+            pow(2, (iNote + ((this->chMIDI[iChannel].iPitchbend - 8192) /
                                       8192.0) + (this->iTranspose / 128) - 9) / 12.0 -
                 (iBlock - 20)) * 440.0 / 32.0 / 50000.0;
-    uint16_t iOPLFNum = (uint16_t)(d + 0.5);
+    uint16_t iOPLFNum = static_cast<uint16_t>(d + 0.5);
     if (iOPLFNum > 1023)
         AdPlug_LogWrite("CMF: This note is out of range! (send this song to "
                         "malvineous@shikadi.net!)\n");
